requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-openai
langchain-core
langgraph
openai
requests
streamlit
pydantic
